# Speed up containsTimestamp