Iterate over string bytes instead of runes to avoid unnecessary utf-8 decoding overhead:
```
goos: linux
goarch: amd64
pkg: github.com/coroot/logparser
                    │    main     │                HEAD                 │
                    │   sec/op    │   sec/op     vs base                │
ContainsTimestamp-8   56.71n ± 2%   42.36n ± 1%  -25.32% (p=0.000 n=10)
```