@AlexanderYastrebov 2025-04-08
Hey folks, today I learned about coroot - its a great idea and implementation :+1: